backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-dotenv>=1.0.0
slowapi>=0.1.8
mistralai>=1.0.0
anthropic>=0.3.0
requests>=2.31.0
pydantic[email]>=2.0.0
python-multipart>=0.0.6
email-validator>=2.0.0